Update codebase to upstream revision c50777a8431157bd4ab3a2c2e2a37e8151ae07da.

+++ drivers/serial/uart_mcux.c
@@ -179,7 +179,7 @@ static int uart_mcux_fifo_fill(const struct device *dev,
 			       int len)
 {
 	const struct uart_mcux_config *config = dev->config;
-	uint8_t num_tx = 0U;
+	int num_tx = 0U;
 
 	while ((len - num_tx > 0) &&
 	       (UART_GetStatusFlags(config->base) & kUART_TxDataRegEmptyFlag)) {
@@ -194,7 +194,7 @@ static int uart_mcux_fifo_read(const struct device *dev, uint8_t *rx_data,
 			       const int len)
 {
 	const struct uart_mcux_config *config = dev->config;
-	uint8_t num_rx = 0U;
+	int num_rx = 0U;
 
 	while ((len - num_rx > 0) &&
 	       (UART_GetStatusFlags(config->base) & kUART_RxDataRegFullFlag)) {
