+++ drivers/serial/uart_mcux_lpsci.c
@@ -89,7 +89,7 @@ static int mcux_lpsci_fifo_fill(const struct device *dev,
 				int len)
 {
 	const struct mcux_lpsci_config *config = dev->config;
-	uint8_t num_tx = 0U;
+	int num_tx = 0U;
 
 	while ((len - num_tx > 0) &&
 	       (LPSCI_GetStatusFlags(config->base)
@@ -105,7 +105,7 @@ static int mcux_lpsci_fifo_read(const struct device *dev, uint8_t *rx_data,
 				const int len)
 {
 	const struct mcux_lpsci_config *config = dev->config;
-	uint8_t num_rx = 0U;
+	int num_rx = 0U;
 
 	while ((len - num_rx > 0) &&
 	       (LPSCI_GetStatusFlags(config->base)
